requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
huggingface-hub

# uncomment to use instructor embeddings
InstructorEmbedding
sentence-transformers